% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn_qk_nono.R
\name{conn_qk}
\alias{conn_qk}
\title{function to integrate db connections}
\usage{
conn_qk(query, db, output)
}
\arguments{
\item{query}{The query, as a string}

\item{db}{The database type: 'dw01', 'dw02', or 'ratt'.}

\item{output}{The type of data returned. Default is to return object as dataframe.
The options are: 'df' (dataframe) or 'dt' (data.table). More to come.}
}
\value{
these functions will return a query
}
\description{
function to integrate db connections
}
