% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_shquote.R
\name{shpaste}
\alias{shpaste}
\title{function to create comma separated list with strings in quotes}
\usage{
shpaste(text, sep)
}
\arguments{
\item{text}{Vector containing the text. Can be in list format as well.}

\item{sep}{The character or string you want to separate the single quoted strings. The default is a comma ','}
}
\value{
shpaste() The function returns one string with the items in a vector or list quoted and separated by a separator
}
\description{
function to create comma separated list with strings in quotes
}
\examples{
ran_vect = runif(n = 10, min = 1, max = 100)
shpaste(ran_vect)
ran_list = lapply(ran_vect, function(x) x)
shpaste(ran_list)
shpaste(ran_list, '--')
}
